Omar - First Integeration Verison
Integration of all components tending to do a randomization for the pieces at the beginning of the scene of the assembly.
All controls are pretty much ok.
We are fixing and handling the snapping mechanics and we are about to finalize the core logic in order to start the polishing phase.