Omar - After Adding Game Logic Script
Game Logic is counting good.
If playr works fire works will fire we need to make the camera stick to the cow front view on order to make the player sees the fireworks.

We can now use the time manager to preview the game time and make logic of winning or loosing.